Prepare Image Optimizer metadata for EdgeZero extraction
